+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - add UI for "format" parameter in "todate" step (not supported for mongo36
   translator)
+- expose `setCodeEditor` to set a custom code editor component for  custom steps
+- custom step which allows users to edit manually a step
 
 ## [0.10.0] - 2020-02-06